Try requestUrl API to get around CORS problems, workaround for now
nostr.build added 'app://obsidian.md' as an origin on their server, so it works for now.

Need to implement requestUrl way for future storage providers as they may not be so helpful.

Added comments with code examples using formData & requestUrl which requires a string or ArrayBuffer as the body of a request.